Extract TimeOfDay value object
